solarsync/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
pydantic==2.9.2
pydantic-settings==2.5.2
apscheduler==3.10.4
httpx==0.27.2
pycryptodome==3.20.0
python-dotenv==1.0.1
requests==2.32.3
numpy==2.1.1
paho-mqtt==2.1.0
